NAV-26033: Lager en egen blob fil for funksjoner relatert til blob. Renamer QueryKeyFactory til et bedre navn.
